backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
openai==1.54.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.24.1
python-jose[cryptography]==3.3.0
supabase==2.3.0
postgrest==0.13.2